requirements.txt
extra_streamlit_components==0.1.60
matplotlib==3.7.2
numpy==1.25.2
opencv_python==4.8.0.76
pandas==2.0.3
Pillow==10.0.1
pydeck==0.8.0
PyJWT==2.8.0
python_bcrypt==0.3.2
PyYAML==6.0.1
scikit_learn==1.3.1
streamlit==1.27.1
streamlit_authenticator==0.2.3
~treamlit==1.24.1
